+++ components/ld2415h/number/sensitivity_number.cpp
@@ -3,7 +3,7 @@
 namespace esphome {
 namespace ld2415h {
 
-void SensitivityNumber::control(uint8_t sensitivity) {
+void SensitivityNumber::control(float sensitivity) {
   this->publish_state(sensitivity);
   this->parent_->set_sensitivity(sensitivity);
 }
